feat(AddArticle): moves article repository protocols to new ArticleRepository file
